+++ iso15118/shared/comm_session.py
@@ -454,20 +454,23 @@ class V2GCommunicationSession(SessionStateMachine):
                         )
                     )
                     return
-            except asyncio.TimeoutError as exc:
-                if self.last_message_sent:
-                    error_msg = (
-                        f"{exc.__class__.__name__} occurred. Waited "
-                        f"for {timeout} s after sending last message: "
-                        f"{str(self.last_message_sent)}"
-                    )
+            except (asyncio.TimeoutError, ConnectionResetError) as exc:
+                if type(exc) == asyncio.TimeoutError:
+                    if self.last_message_sent:
+                        error_msg = (
+                            f"{exc.__class__.__name__} occurred. Waited "
+                            f"for {timeout} s after sending last message: "
+                            f"{str(self.last_message_sent)}"
+                        )
+                    else:
+                        error_msg = (
+                            f"{exc.__class__.__name__} occurred. Waited "
+                            f"for {timeout} s. No V2GTP message was "
+                            "previously sent. This is probably a timeout "
+                            f"while waiting for SupportedAppProtocolReq"
+                        )
                 else:
-                    error_msg = (
-                        f"{exc.__class__.__name__} occurred. Waited "
-                        f"for {timeout} s. No V2GTP message was "
-                        "previously sent. This is probably a timeout "
-                        f"while waiting for SupportedAppProtocolReq"
-                    )
+                    error_msg = f"{exc.__class__.__name__} occurred. {str(exc)}"
 
                 self.stop_reason = StopNotification(False, error_msg, self.peer_name)