+++ iso15118/shared/messages/datatypes.py
@@ -52,7 +52,7 @@ class PhysicalValue(BaseModel):
             )
         return values
 
-    def get_physical_value(self) -> float:
+    def get_decimal_value(self) -> float:
         return self.value * 10**self.multiplier